requirements.txt
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
boto3==1.17.109
botocore==1.20.109
cffi==1.14.6
click==8.0.1
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
passlib==1.7.4
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2022.1
s3transfer==0.4.2
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
